refactors the config manager into clearer modules
