+++ agent/configmgr/fleet/mocks.go
@@ -1,0 +1,88 @@
+package fleet
+
+import (
+	"context"
+	"encoding/base64"
+	"encoding/json"
+	"log/slog"
+	"time"
+
+	"github.com/stretchr/testify/mock"
+
+	"github.com/netboxlabs/orb-agent/agent/backend"
+	"github.com/netboxlabs/orb-agent/agent/config"
+	"github.com/netboxlabs/orb-agent/agent/policies"
+)
+
+// RawJWTWithClaims builds a raw JWT string with the given claims.
+// Signature is a dummy value; ParseUnverified only inspects header/payload.
+func RawJWTWithClaims(claims map[string]any) string {
+	header := map[string]any{
+		"alg": "RS256",
+		"kid": "test-key",
+		"typ": "JWT",
+	}
+	h, _ := json.Marshal(header)
+	p, _ := json.Marshal(claims)
+	return base64.RawURLEncoding.EncodeToString(h) + "." + base64.RawURLEncoding.EncodeToString(p) + ".sig"
+}
+
+// mockBackend implements the Backend interface for testing
+type mockBackend struct {
+	mock.Mock
+}
+
+func (m *mockBackend) Configure(logger *slog.Logger, repo policies.PolicyRepo, config map[string]any, commons config.BackendCommons) error {
+	args := m.Called(logger, repo, config, commons)
+	return args.Error(0)
+}
+
+func (m *mockBackend) Version() (string, error) {
+	args := m.Called()
+	return args.String(0), args.Error(1)
+}
+
+func (m *mockBackend) Start(ctx context.Context, cancelFunc context.CancelFunc) error {
+	args := m.Called(ctx, cancelFunc)
+	return args.Error(0)
+}
+
+func (m *mockBackend) Stop(ctx context.Context) error {
+	args := m.Called(ctx)
+	return args.Error(0)
+}
+
+func (m *mockBackend) FullReset(ctx context.Context) error {
+	args := m.Called(ctx)
+	return args.Error(0)
+}
+
+func (m *mockBackend) GetStartTime() time.Time {
+	args := m.Called()
+	return args.Get(0).(time.Time)
+}
+
+func (m *mockBackend) GetCapabilities() (map[string]any, error) {
+	args := m.Called()
+	return args.Get(0).(map[string]any), args.Error(1)
+}
+
+func (m *mockBackend) GetRunningStatus() (backend.RunningStatus, string, error) {
+	args := m.Called()
+	return args.Get(0).(backend.RunningStatus), args.String(1), args.Error(2)
+}
+
+func (m *mockBackend) GetInitialState() backend.RunningStatus {
+	args := m.Called()
+	return args.Get(0).(backend.RunningStatus)
+}
+
+func (m *mockBackend) ApplyPolicy(data policies.PolicyData, updatePolicy bool) error {
+	args := m.Called(data, updatePolicy)
+	return args.Error(0)
+}
+
+func (m *mockBackend) RemovePolicy(data policies.PolicyData) error {
+	args := m.Called(data)
+	return args.Error(0)
+}
